+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+### Fixed
+
+- Deprecation issues in PHP 8.4 with fputcsv.
+- Various code quality, linting, and whitespace issues using PhpStan, PHP CS Fixer, and Rector.
+
 ## [3.0.3] - 2025-04-05
 
 ### Fixed
@@ -419,6 +426,7 @@
 
 - Public release of the Local Geohistory Project: Application repository.
 
+[Unreleased]: https://github.com/localgeohistoryproject/application/compare/v3.0.3...develop
 [3.0.3]: https://github.com/localgeohistoryproject/application/compare/v3.0.2...v3.0.3
 [3.0.2]: https://github.com/localgeohistoryproject/application/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/localgeohistoryproject/application/compare/v3.0.0...v3.0.1